[Update] Fixed the issue in v0.15.1 where the Kunlun plugin failed to initialize due to circular references (#201)
